+++ samples/nrf9160/lwm2m_client/src/lwm2m/lwm2m_firmware.c
@@ -6,13 +6,13 @@
 
 #include <zephyr.h>
 #include <stdio.h>
-#include <bsd.h>
+#include <nrf_modem.h>
 #include <drivers/flash.h>
 #include <dfu/dfu_target.h>
 #include <dfu/mcuboot.h>
 #include <logging/log_ctrl.h>
 #include <net/lwm2m.h>
-#include <modem/bsdlib.h>
+#include <modem/nrf_modem_lib.h>
 #include <power/reboot.h>
 
 #include "settings.h"
@@ -192,7 +192,7 @@ int lwm2m_init_firmware(void)
 
 void lwm2m_verify_modem_fw_update(void)
 {
-	int ret = bsdlib_get_init_ret();
+	int ret = nrf_modem_lib_get_init_ret();
 	struct update_counter counter;
 
 	/* Handle return values relating to modem firmware update */